requirements.txt
Flask==2.2.3
langchain==0.1.13
langchain-openai==0.1.0
transformers==4.39.1
torch==2.2.1
Pillow==9.5.0
youtube_search==2.1.2
python-dotenv==1.0.1
Werkzeug==2.2.2
timm==0.9.16